+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Improve ASM engine soundness to better match WALA edge sets (#45)
+
 ### Security
 
 ## [v0.1.0] - 2026-03-22